+++ react/components/loaders/AddressCreateLoading.js
@@ -1,30 +1,14 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import {
-  SkeletonPiece,
-  SkeletonBox,
-  BaseLoading,
-} from 'vtex.my-account-commons'
+import { BaseLoading } from 'vtex.my-account-commons'
 
 import { headerConfig } from '../pages/AddressCreate'
+import FormLoader from './FormLoader'
 
 const AddressCreateLoading = ({ data }) => {
   return (
     <BaseLoading queryData={data} headerConfig={headerConfig()}>
-      <SkeletonBox shouldAllowGrowing maxWidthStep={6}>
-        <div className="mb8 mt4">
-          <SkeletonPiece width={100} />
-          <SkeletonPiece width={100} />
-        </div>
-        <div className="mb9">
-          <SkeletonPiece width={100} />
-          <SkeletonPiece width={100} />
-        </div>
-        <div className="mb8">
-          <SkeletonPiece width={40} />
-          <SkeletonPiece width={40} />
-        </div>
-      </SkeletonBox>
+      <FormLoader />
     </BaseLoading>
   )
 }